+++ src/main/java/com/foxminded/counter/Counter.java
@@ -4,5 +4,5 @@ import java.util.Map;
 
 @FunctionalInterface
 public interface Counter<T> {
-    Map<T, Integer> count(String inputString);
+    Map<T, Long> count(String inputString);
 }
